feat: students can view their group from the class page
