+++ client/src/components/AddUserDialog/index.js
@@ -1,0 +1,2 @@
+import AddUserDialog from "./AddUserDialog";
+export default AddUserDialog;
