+++ client/src/components/UserList/UserList.js
@@ -1,19 +1,8 @@
-import { useState } from "react";
-import {
-  Button,
-  Dialog,
-  DialogActions,
-  DialogContent,
-  DialogContentText,
-  DialogTitle,
-  TextField,
-  Select,
-  MenuItem,
-  makeStyles,
-} from "@material-ui/core";
-import AddIcon from "@material-ui/icons/Add";
+import { makeStyles } from "@material-ui/core";
 import { useParams } from "react-router-dom";
 import axios from "axios";
+
+import AddUserDialog from "../AddUserDialog";
 
 const useStyles = makeStyles({
   tableHeader: {
@@ -38,14 +27,9 @@ const useStyles = makeStyles({
 
 function UserList(props) {
   const { classID } = useParams();
-  // Form values
-  const [userEmails, setUserEmails] = useState("");
-  const [newUserRole, setNewUserRole] = useState("STUDENT");
 
   // Queried values
   const { curUserRole, queriedUserList, refreshClassData } = props;
-
-  const [dialogOpen, setDialogOpen] = useState(false);
 
   const styles = useStyles();
 
@@ -56,20 +40,7 @@ function UserList(props) {
     );
   }, []);
 
-  const handleDialogOpen = () => {
-    setDialogOpen(true);
-  };
-
-  const handleDialogClose = () => {
-    setDialogOpen(false);
-  };
-
-  const handleSubmit = (event) => {
-    event.preventDefault();
-    handleAddUsers(userEmails, newUserRole);
-  };
-
-  // Current only handles adding 1 at a time, and only students
+  // Current only handles adding 1 at a time
   const handleAddUsers = (userEmails, newUserRole) => {
     axios
       .post(
@@ -82,8 +53,6 @@ function UserList(props) {
       )
       .then((response) => {
         console.log(response.data);
-        setUserEmails("");
-        handleDialogClose();
       })
       .then(() => refreshClassData(classID))
       .catch((err) => console.log(err));
@@ -96,40 +65,7 @@ function UserList(props) {
         {
           // If user is a mentor, render the add users button
           curUserRole === "STUDENT" || (
-            <>
-              <Button className={styles.button} onClick={handleDialogOpen}>
-                <AddIcon />
-              </Button>
-              <Dialog open={dialogOpen} onClose={handleDialogClose}>
-                <DialogTitle id="form-dialog-title">Add Users</DialogTitle>
-                <DialogContent>
-                  <DialogContentText>
-                    Add user(s) by typing in their emails.
-                  </DialogContentText>
-                  <TextField
-                    autoFocus
-                    id="email"
-                    label="Email Address"
-                    type="email"
-                    fullWidth
-                    value={userEmails}
-                    onChange={(event) => setUserEmails(event.target.value)}
-                  />
-                  <Select
-                    value={newUserRole}
-                    onChange={(event) => setNewUserRole(event.target.value)}
-                    label="New User Role"
-                  >
-                    <MenuItem value={"STUDENT"}>Student</MenuItem>
-                    <MenuItem value={"MENTOR"}>Mentor</MenuItem>
-                  </Select>
-                </DialogContent>
-                <DialogActions>
-                  <Button onClick={handleDialogClose}>Cancel</Button>
-                  <Button onClick={handleSubmit}>Add</Button>
-                </DialogActions>
-              </Dialog>
-            </>
+            <AddUserDialog handleAddUsers={handleAddUsers} />
           )
         }
       </div>
